requirements.txt
gradio
transformers
torch
faiss-cpu
opencv-python
pillow
